rag_core/requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
chromadb
pypdf
python-dotenv
langchain-google-genai
psycopg2-binary
langchain-postgres
fastapi
uvicorn
python-multipart
unstructured
markdown
openpyxl
python-docx
python-pptx
olefile
networkx
